chores: format code, meson setup

+++ TME8/src/ServerSocket.h
@@ -6,19 +6,20 @@
 namespace pr {
 
 class ServerSocket {
-	int socketfd;
+  int socketfd;
 
-public :
-	// Demarre l'ecoute sur le port donne
-	ServerSocket(int port);
+public:
+  // Demarre l'ecoute sur le port donne
+  ServerSocket(short port);
 
-	int getFD() { return socketfd;}
-	bool isOpen() const {return socketfd != -1;}
+  int getFD() { return socketfd; }
+  bool isOpen() const { return socketfd != -1; }
 
-	Socket accept();
+  Socket accept();
 
-	void close();
+  void close();
 };
 
-} // ns pr
+} // namespace pr
+
 #endif /* SRC_SERVERSOCKET_H_ */
