+++ TME8/src/TCPServer.h
@@ -1,26 +1,26 @@
 #ifndef SRC_TCPSERVER_H_
 #define SRC_TCPSERVER_H_
 
-#include <thread>
-#include "ServerSocket.h"
 #include "ConnectionHandler.h"
+#include "ServerSocket.h"
+#include <thread>
 
 namespace pr {
 
 // un serveur TCP, la gestion des connections est déléguée
 class TCPServer {
-	ServerSocket * ss; // la socket d'attente si elle est instanciee
-	ConnectionHandler * handler; // le gestionnaire de session passe a la constru
-	// a completer
-public :
-	TCPServer(ConnectionHandler * handler): ss(nullptr),handler(handler) {}
-	// Tente de creer une socket d'attente sur le port donné
-	bool startServer (int port);
+  ServerSocket *ss;           // la socket d'attente si elle est instanciee
+  ConnectionHandler *handler; // le gestionnaire de session passe a la constru
+                              // a completer
+public:
+  TCPServer(ConnectionHandler *handler) : ss(nullptr), handler(handler) {}
+  // Tente de creer une socket d'attente sur le port donné
+  bool startServer(short port);
 
-	// stoppe le serveur
-	void stopServer () ;
+  // stoppe le serveur
+  void stopServer();
 };
 
-} // ns pr
+} // namespace pr
 
 #endif /* SRC_TCPSERVER_H_ */
